+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Can pass schema to Python class as well as filename
+
 ## [0.3.0] - 2021-01-27
 
 ### Added